R6 Class representing an Export

R6 Class representing a resource for managing volume export jobs.

Super class

sevenbridges2::Item -> Export

Public fields

URL

List of URL endpoints for this resource.

id

Export job string identifier.

state

The state of the export job. Possible values are:

PENDING: the export is queued;

RUNNING: the export is running;

COMPLETED: the export has completed successfully;

FAILED: the export has failed.

source

List containing source file id that is being exported to the volume.

destination

List containing destination volume id and location (file name) on the volume where the file is being exported.

overwrite

Whether the exported file name was overwritten or not, if another one with the same name had already existed on the volume.

started_on

Time when the export job started.

finished_on

Time when the export job ended.

properties

List of volume properties set.

error

In case of error in the export job, standard API error is returned here.

result

File object that was exported.

Methods

Public methods

Export$new()

Export$print()

Export$reload()

Export$clone()

Method new()

Create a new Export object.

Usage

Export$new(res = NA, ...)

Arguments

res

Response containing Export job information.

...

Other response arguments.

Method print()

Print method for Export class.

Usage

Export$print()

Examples

\dontrun{
# x is API response when export is requested
export_object <- Export$new(
                    res = x,
                    href = x$href,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Print export object
 export_object$print()
}

Method reload()

Reload Export object information.

Usage

Export$reload(...)

Arguments

...

Other arguments that can be passed to core api() function like 'fields', etc.

Returns

Export object.

Examples

\dontrun{
# x is API response when export is requested
export_object <- Export$new(
                    res = x,
                    href = x$href,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Reload export object
 export_object$reload()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

Export$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Export$print`
## ------------------------------------------------

if (FALSE) {
# x is API response when export is requested
export_object <- Export$new(
                    res = x,
                    href = x$href,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Print export object
 export_object$print()
}

## ------------------------------------------------
## Method `Export$reload`
## ------------------------------------------------

if (FALSE) {
# x is API response when export is requested
export_object <- Export$new(
                    res = x,
                    href = x$href,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Reload export object
 export_object$reload()
}