R6 Class Representing a Rate Limit for a user

Rate object containing information about user's rate limit.

Details

This is the main object for Rate Limit.

Super class

sevenbridges2::Item -> Rate

Public fields

rate

A list containing the information about user's current rate limit. It consists of the following fields:

limit Indicates how many requests can be made in five minutes.

remaining Indicates how many requests remain.

reset Indicates the time when the request rate limit will be reset.

instance

A list containing the information about user's current instance limit. It consists of the following fields:

limit Indicates the total number of instances available to the user. For the first few months, instance limits are unlimited. This is indicated by a special limit of -1. Correspondingly, the remaining value is high.

remaining Indicates the number of the instances that are available at the moment. For the first few months, instance limits are unlimited. This is indicated by a high remaining value. Correspondingly, the limit is set to a special value of -1.

Methods

Public methods

Rate$new()

Rate$print()

Rate$clone()

Inherited methods

sevenbridges2::Item$reload()

Method new()

Create a new Rate limit object.

Usage

Rate$new(res = NA, ...)

Arguments

res

Response containing Rate limit object info.

...

Other response arguments.

Method print()

Print rate limit information as a bullet list.

Usage

Rate$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

Rate$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.