R6 Class Representing a platform User

User object containing user information.

Details

This is main object for Users.

Super class

sevenbridges2::Item -> User

Public fields

URL

List of URL endpoints for this resource.

username

User name.

email

User's email address.

first_name

User's first name.

last_name

User's last name.

affiliation

The company or the institute the user is affiliated with.

phone

User's phone number.

address

User's residential address.

city

User's city of residence.

state

User's state of residence.

country

User's country of residence.

zip_code

Zip code for the user's residence.

role

User's role.

tags

Platform tags associated with the user.

Methods

Public methods

User$new()

User$print()

User$reload()

User$clone()

Method new()

Create a new User object.

Usage

User$new(res = NA, ...)

Arguments

res

Response containing User object information.

...

Other response arguments.

Method print()

Print user information as bullet list.

Usage

User$print()

Method reload()

Reload User object information.

Usage

User$reload(...)

Arguments

...

Other arguments that can be passed to core api() function like 'fields', etc.

Returns

User object.

Method clone()

The objects of this class are cloneable with this method.

Usage

User$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.