R6 Class representing a VolumeContentCollection

R6 Class representing a resource for managing volume content collections.

Super class

sevenbridges2::Collection -> VolumeContentCollection

Public fields

prefixes

Prefixes on the volume, returned in API response.

Methods

Public methods

VolumeContentCollection$new()

VolumeContentCollection$print()

VolumeContentCollection$next_page()

VolumeContentCollection$prev_page()

VolumeContentCollection$all()

VolumeContentCollection$clone()

Method new()

Create new VolumeContentCollection object.

Usage

VolumeContentCollection$new(res = NA, ...)

Arguments

res

Response containing VolumeContentCollection object fields.

...

Other response arguments.

Method print()

Print method for VolumeContentCollection class.

Usage

VolumeContentCollection$print(n = 10)

Arguments

n

Number of items to print in console.

Examples

\dontrun{
# x is API response when volume content collection is requested
vol_con_col_object <- VolumeContentCollection$new(
                    res = x,
                    href = x$href,
                    links = x$links,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Print volume content collection object
 vol_con_col_object$print()
}

Method next_page()

Return next page of results.

Usage

VolumeContentCollection$next_page(...)

Arguments

...

Other arguments or query parameters that can be passed to the core api() function like 'advance_access', 'fields' etc.

Examples

\dontrun{
# x is API response when volume content collection is requested
vol_con_col_object <- VolumeContentCollection$new(
                    res = x,
                    href = x$href,
                    links = x$links,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Get next page of results
 vol_con_col_object$next_page()
}

Method prev_page()

Return the previous page of results.

Usage

VolumeContentCollection$prev_page()

Examples

\dontrun{
# x is API response when volume content collection is requested
vol_con_col_object <- VolumeContentCollection$new(
                    res = x,
                    href = x$href,
                    links = x$links,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Get previous page of results
 vol_con_col_object$prev_page()
}

Method all()

Fetches all available items.

Usage

VolumeContentCollection$all(...)

Arguments

...

Other arguments or query parameters that can be passed to the core api() function like 'advance_access', 'fields' etc.

Examples

\dontrun{
# x is API response when volume content collection is requested
vol_con_col_object <- VolumeContentCollection$new(
                    res = x,
                    href = x$href,
                    links = x$links,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Get all results
 vol_con_col_object$all()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

VolumeContentCollection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `VolumeContentCollection$print`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# x is API response when volume content collection is requested
vol_con_col_object <- VolumeContentCollection$new(
                    res = x,
                    href = x$href,
                    links = x$links,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Print volume content collection object
 vol_con_col_object$print()
} # }

## ------------------------------------------------
## Method `VolumeContentCollection$next_page`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# x is API response when volume content collection is requested
vol_con_col_object <- VolumeContentCollection$new(
                    res = x,
                    href = x$href,
                    links = x$links,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Get next page of results
 vol_con_col_object$next_page()
} # }

## ------------------------------------------------
## Method `VolumeContentCollection$prev_page`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# x is API response when volume content collection is requested
vol_con_col_object <- VolumeContentCollection$new(
                    res = x,
                    href = x$href,
                    links = x$links,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Get previous page of results
 vol_con_col_object$prev_page()
} # }

## ------------------------------------------------
## Method `VolumeContentCollection$all`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# x is API response when volume content collection is requested
vol_con_col_object <- VolumeContentCollection$new(
                    res = x,
                    href = x$href,
                    links = x$links,
                    auth = auth,
                    response = attr(x, "response")
                   )

 # Get all results
 vol_con_col_object$all()
} # }