R6 Class representing a VolumePrefix

R6 Class representing a resource for managing VolumePrefix objects.

Super class

sevenbridges2::Item -> VolumePrefix

Public fields

URL

List of URL endpoints for this resource.

prefix

File/prefix name on the volume.

volume

Volume id.

Methods

Public methods

VolumePrefix$new()

VolumePrefix$print()

VolumePrefix$reload()

VolumePrefix$list_contents()

VolumePrefix$import()

VolumePrefix$clone()

Method new()

Create a new VolumePrefix object.

Usage

VolumePrefix$new(res = NA, ...)

Arguments

res

Response containing VolumePrefix object info.

...

Other response arguments.

Method print()

Print method for VolumePrefix class.

Usage

VolumePrefix$print()

Method reload()

Reload VolumePrefix object information.

Usage

VolumePrefix$reload()

Method list_contents()

List volume folder contents. This call lists the contents of a specific volume folder.

Usage

VolumePrefix$list_contents(
  limit = getOption("sevenbridges2")$limit,
  continuation_token = NULL,
  ...
)

Arguments

limit

The maximum number of collection items to return for a single request. Minimum value is 1. The maximum value is 100 and the default value is 50. This is a pagination-specific attribute.

continuation_token

Continuation token received to use for next chunk of results. Behaves similarly like offset parameter.

...

Other arguments that can be passed to core api() function, like 'fields' for example. With fields parameter you can specify a subset of fields to include in the response. You can use: href, location, volume, type, metadata, _all. Default: _all.

Returns

VolumeContentCollection object containing list of VolumeFile and VolumePrefix objects.

Method import()

This call lets you queue a job to import this file or folder from a volume into a project on the Platform.
Essentially, you are importing an item from your cloud storage provider (Amazon Web Services, Google Cloud Storage, Azure or Ali Cloud) via the volume onto the Platform.
If successful, an alias will be created on the Platform. Aliases appear on the Platform and can be copied, executed, and modified as such. They refer back to the respective item on the given volume.

Usage

VolumePrefix$import(
  destination_project = NULL,
  destination_parent = NULL,
  name = NULL,
  overwrite = FALSE,
  autorename = FALSE,
  preserve_folder_structure = NULL,
  ...
)

Arguments

destination_project

String destination project id or Project object. Not required, but either destination_project or destination_parent directory must be provided.

destination_parent

String folder id or File object (with type = 'FOLDER'). Not required, but either destination_project or destination_parent directory must be provided.

name

The name of the alias to create. This name should be unique to the project. If the name is already in use in the project, you should use the overwrite query parameter in this call to force any item with that name to be deleted before the alias is created. If name is omitted, the alias name will default to the last segment of the complete location (including the prefix) on the volume.

Segments are considered to be separated with forward slashes /. Allowed characters in file names are all alphanumeric and special characters except forward slash /, while folder names can contain alphanumeric and special characters _, - and ..

overwrite

Set to TRUE if you want to overwrite the item if another one with the same name already exists at the destination. Bear in mind that if used with folders import, the folder's content (files with the same name) will be overwritten, not the whole folder.

autorename

Set to TRUE if you want to automatically rename the item (by prefixing its name with an underscore and number) if another one with the same name already exists at the destination. Bear in mind that if used with folders import, the folder content will be renamed, not the whole folder.

preserve_folder_structure

Set to TRUE if you want to keep the exact source folder structure. The default value is TRUE if the item being imported is a folder. Should not be used if you are importing a file. Bear in mind that if you use preserve_folder_structure = FALSE, that the response will be the parent folder object containing imported files alongside with other files if they exist.

...

Other arguments that can be passed to core api() function like 'fields', etc.

Returns

Import object.

Method clone()

The objects of this class are cloneable with this method.

Usage

VolumePrefix$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.